+++ CHANGES.md
@@ -4,17 +4,18 @@
 ## Version 0.4.0
 *Dev*
 
-**PetroFit Improvements**
+**PetroFit Enhancements**
 
 - Removed outdated `petrofit.utils.cutout` function. [:pr:#93]
 - `model_to_image` function upgraded to use `astropy.convolution.utils.discretize_model`. [:pr:#95]
 - `PSFModel` has been upgraded to `PSFConvolvedImageModel`.`PSFConvolvedImageModel` uses `__init__` function to wrap models. `PSFModel` class has been deprecated and will be removed in `v0.5`. [:pr:#100]
 - An `origin` parameter has been added to `petrofit.models.make_grid`. [:pr:#100]
+- `fitting.py` and `models.py` have been moved into a new `modeling` folder. This allows for all modeling functions and classes to be imported as `from petrofit.modeling import ...`. [:pr:#100] 
 
 ## Version 0.3.1
 *March 8th 2022*
 
-**PetroFit Improvements**
+**PetroFit Enhancements**
 
 - Removed kcorrect from Docker file. [:pr:#77]
 - Added badges to `README.rst`.